// src/modules/admin/types/admin.types.ts
export enum AdminErrorCode {
  ORGANIZATION_NOT_FOUND = 'ORGANIZATION_NOT_FOUND',
  MEMBER_NOT_FOUND = 'MEMBER_NOT_FOUND',
  CANNOT_REMOVE_LAST_ADMIN = 'CANNOT_REMOVE_LAST_ADMIN',
  CANNOT_GRANT_ADMIN = 'CANNOT_GRANT_ADMIN',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  CUSTOMER_NOT_FOUND = 'CUSTOMER_NOT_FOUND',
  PROVIDER_NOT_FOUND = 'PROVIDER_NOT_FOUND',
  DELIVERY_PARTNER_NOT_FOUND = 'DELIVERY_PARTNER_NOT_FOUND',
  BOOKING_NOT_FOUND = 'BOOKING_NOT_FOUND',
  ASSIGNMENT_NOT_FOUND = 'ASSIGNMENT_NOT_FOUND',
  SERVICE_NOT_FOUND = 'SERVICE_NOT_FOUND',
  CATEGORY_NOT_FOUND = 'CATEGORY_NOT_FOUND',
  PAYMENT_NOT_FOUND = 'PAYMENT_NOT_FOUND',
  REFUND_NOT_FOUND = 'REFUND_NOT_FOUND',
  TRANSACTION_NOT_FOUND = 'TRANSACTION_NOT_FOUND',
  EARNING_NOT_FOUND = 'EARNING_NOT_FOUND',
  REVIEW_NOT_FOUND = 'REVIEW_NOT_FOUND',
  NOTIFICATION_NOT_FOUND = 'NOTIFICATION_NOT_FOUND',
  COMMUNICATION_NOT_FOUND = 'COMMUNICATION_NOT_FOUND',
  SUPPORT_TICKET_NOT_FOUND = 'SUPPORT_TICKET_NOT_FOUND',
  DISPUTE_NOT_FOUND = 'DISPUTE_NOT_FOUND',
  AUDIT_EVENT_NOT_FOUND = 'AUDIT_EVENT_NOT_FOUND',
  TENANT_MISMATCH = 'TENANT_MISMATCH',
  FORBIDDEN_PRIVILEGE_ESCALATION = 'FORBIDDEN_PRIVILEGE_ESCALATION',
  INVALID_OPERATION_STATE = 'INVALID_OPERATION_STATE',
  ASSIGNMENT_INVALID_STATE = 'ASSIGNMENT_INVALID_STATE',
  BOOKING_INVALID_STATE_TRANSITION = 'BOOKING_INVALID_STATE_TRANSITION',
  INVALID_ROLE = 'INVALID_ROLE',
  MEMBER_ALREADY_EXISTS = 'MEMBER_ALREADY_EXISTS',
  CONFLICT = 'CONFLICT',
  VALIDATION_ERROR = 'VALIDATION_ERROR',
}

export enum AdminAuditEvent {
  ORG_SETTINGS_UPDATED = 'ORG_SETTINGS_UPDATED',
  ORGANIZATION_UPDATED = 'ORGANIZATION_UPDATED',
  MEMBER_INVITED = 'MEMBER_INVITED',
  MEMBER_CREATED = 'MEMBER_CREATED',
  MEMBER_UPDATED = 'MEMBER_UPDATED',
  MEMBER_STATUS_CHANGED = 'MEMBER_STATUS_CHANGED',
  MEMBER_ROLE_UPDATED = 'MEMBER_ROLE_UPDATED',
  MEMBER_ROLE_CHANGED = 'MEMBER_ROLE_CHANGED',
  MEMBER_REMOVED = 'MEMBER_REMOVED',
  MEMBER_ACTIVATED = 'MEMBER_ACTIVATED',
  MEMBER_DEACTIVATED = 'MEMBER_DEACTIVATED',
  MEMBER_SUSPENDED = 'MEMBER_SUSPENDED',
  USER_STATUS_CHANGED = 'USER_STATUS_CHANGED',
  USER_ACTIVATED = 'USER_ACTIVATED',
  USER_DEACTIVATED = 'USER_DEACTIVATED',
  USER_SUSPENDED = 'USER_SUSPENDED',
  USER_SESSIONS_REVOKED = 'USER_SESSIONS_REVOKED',
  CUSTOMER_UPDATED = 'CUSTOMER_UPDATED',
  CUSTOMER_SUSPENDED = 'CUSTOMER_SUSPENDED',
  CUSTOMER_ACTIVATED = 'CUSTOMER_ACTIVATED',
  PROVIDER_UPDATED = 'PROVIDER_UPDATED',
  PROVIDER_VERIFIED = 'PROVIDER_VERIFIED',
  PROVIDER_ACTIVATED = 'PROVIDER_ACTIVATED',
  PROVIDER_SUSPENDED = 'PROVIDER_SUSPENDED',
  DELIVERY_PARTNER_UPDATED = 'DELIVERY_PARTNER_UPDATED',
  DELIVERY_PARTNER_VERIFIED = 'DELIVERY_PARTNER_VERIFIED',
  DELIVERY_PARTNER_ACTIVATED = 'DELIVERY_PARTNER_ACTIVATED',
  DELIVERY_PARTNER_SUSPENDED = 'DELIVERY_PARTNER_SUSPENDED',
  BOOKING_CANCELLED = 'BOOKING_CANCELLED',
  BOOKING_CANCELLED_BY_ADMIN = 'BOOKING_CANCELLED_BY_ADMIN',
  BOOKING_RESCHEDULED = 'BOOKING_RESCHEDULED',
  BOOKING_RESCHEDULED_BY_ADMIN = 'BOOKING_RESCHEDULED_BY_ADMIN',
  BOOKING_NOTE_ADDED = 'BOOKING_NOTE_ADDED',
  BOOKING_NOTE_ADDED_BY_ADMIN = 'BOOKING_NOTE_ADDED_BY_ADMIN',
  ASSIGNMENT_REASSIGNED = 'ASSIGNMENT_REASSIGNED',
  ASSIGNMENT_CANCELLED = 'ASSIGNMENT_CANCELLED',
  CATEGORY_CREATED = 'CATEGORY_CREATED',
  CATEGORY_UPDATED = 'CATEGORY_UPDATED',
  CATEGORY_DELETED = 'CATEGORY_DELETED',
  SERVICE_CREATED = 'SERVICE_CREATED',
  SERVICE_UPDATED = 'SERVICE_UPDATED',
  SERVICE_PUBLISHED = 'SERVICE_PUBLISHED',
  SERVICE_UNPUBLISHED = 'SERVICE_UNPUBLISHED',
  CATALOG_ADMIN_ACTION = 'CATALOG_ADMIN_ACTION',
  PAYMENT_OVERRIDDEN = 'PAYMENT_OVERRIDDEN',
  PAYMENT_ADMIN_ACTION = 'PAYMENT_ADMIN_ACTION',
  REFUND_PROCESSED = 'REFUND_PROCESSED',
  REFUND_FAILED = 'REFUND_FAILED',
  REFUND_ADMIN_ACTION = 'REFUND_ADMIN_ACTION',
  REVIEW_MODERATED = 'REVIEW_MODERATED',
  REVIEW_REPORT_RESOLVED = 'REVIEW_REPORT_RESOLVED',
  REVIEW_MODERATION_ACTION = 'REVIEW_MODERATION_ACTION',
  BROADCAST_SENT = 'BROADCAST_SENT',
  NOTIFICATION_BROADCAST = 'NOTIFICATION_BROADCAST',
  NOTIFICATION_TEMPLATE_UPDATED = 'NOTIFICATION_TEMPLATE_UPDATED',
  SUPPORT_TICKET_ASSIGNED = 'SUPPORT_TICKET_ASSIGNED',
  SUPPORT_TICKET_ESCALATED = 'SUPPORT_TICKET_ESCALATED',
  SUPPORT_TICKET_RESOLVED = 'SUPPORT_TICKET_RESOLVED',
  SUPPORT_TICKET_CLOSED = 'SUPPORT_TICKET_CLOSED',
  SUPPORT_PRIORITY_UPDATED = 'SUPPORT_PRIORITY_UPDATED',
  SUPPORT_NOTE_ADDED = 'SUPPORT_NOTE_ADDED',
  DISPUTE_CREATED = 'DISPUTE_CREATED',
  DISPUTE_ASSIGNED = 'DISPUTE_ASSIGNED',
  DISPUTE_ESCALATED = 'DISPUTE_ESCALATED',
  DISPUTE_RESOLVED = 'DISPUTE_RESOLVED',
  DISPUTE_REJECTED = 'DISPUTE_REJECTED',
  DISPUTE_PRIORITY_UPDATED = 'DISPUTE_PRIORITY_UPDATED',
}